Bug-1254567 - Add footer view to menu view & use rounded toolbar for top right and top left
We're going to have to think about the footer view openMenuImage. Right now I'm just adding it, but I am unsure whether it is the right thing to do having it there. A lot of that will depend on what happens when the menu animations come through.
Also, if we decide to leave it there, the position of the menu button will change so hardcoding it's location to the center of the view is not appropriate.  This will probably be resolved when we figure out the API for presenting the menu.
Like the toolbar, the footer view may not be displayed - in this instance iPhone landscape and iPad sizes. This will also need to be dealt with later.